+++ packages/asset-swapper/contracts/src/KyberDmmSampler.sol
@@ -22,12 +22,21 @@ pragma experimental ABIEncoderV2;
 
 import "@0x/contracts-zero-ex/contracts/src/transformers/bridges/mixins/MixinKyberDmm.sol";
 import "./SwapRevertSampler.sol";
-interface IKyberDmmFactory {
 
-    function getPoolAtIndex(address token0, address token1, uint256 index)
+interface IKyberDmmPool {
+
+    function totalSupply()
         external
         view
-        returns (address);
+        returns (uint256);
+}
+
+interface IKyberDmmFactory {
+
+    function getPools(address token0, address token1)
+        external
+        view
+        returns (address[] memory _tokenPools);
 }
 
 
@@ -140,17 +149,26 @@ contract KyberDmmSampler is
         view
         returns (address[] memory pools)
     {
-        pools = new address[](path.length - 1);
         IKyberDmmFactory factory = IKyberDmmFactory(IKyberDmmRouter(router).factory());
+        pools = new address[](path.length - 1);
         for (uint256 i = 0; i < pools.length; i++) {
-            // Currently only supporting the first pool found at the index
+            // find the best pool
+            address[] memory allPools;
             try
-                factory.getPoolAtIndex
+                factory.getPools
                     {gas: KYBER_DMM_CALL_GAS}
-                    (path[i], path[i + 1], 0)
-                returns (address pool)
+                    (path[i], path[i + 1])
+                returns (address[] memory allPools)
             {
-                pools[i] = pool;
+                uint256 maxSupply = 0;
+                require(allPools.length >= 1, "KyberDMMSampler/NO_POOLS_FOUND");
+                for (uint256 j = 0; j < allPools.length; j++) {
+                    uint256 totalSupply = IKyberDmmPool(allPools[j]).totalSupply();
+                    if (totalSupply > maxSupply) {
+                        maxSupply = totalSupply;
+                        pools[i] = allPools[j];
+                    }
+                }
             } catch (bytes memory) {
                 return new address[](0);
             }