+++ packages/asset-swapper/src/utils/market_operation_utils/bridge_source_utils.ts
@@ -18,6 +18,7 @@ import {
     ELLIPSIS_BSC_INFOS,
     FIREBIRDONESWAP_BSC_INFOS,
     FIREBIRDONESWAP_POLYGON_INFOS,
+    JETSWAP_ROUTER_BY_CHAIN_ID,
     JULSWAP_ROUTER_BY_CHAIN_ID,
     KYBER_BANNED_RESERVES,
     KYBER_BRIDGED_LIQUIDITY_PREFIX,
@@ -32,6 +33,7 @@ import {
     QUICKSWAP_ROUTER_BY_CHAIN_ID,
     SADDLE_MAINNET_INFOS,
     SHELL_POOLS_BY_CHAIN_ID,
+    SHIBASWAP_ROUTER_BY_CHAIN_ID,
     SMOOTHY_BSC_INFOS,
     SMOOTHY_MAINNET_INFOS,
     SNOWSWAP_MAINNET_INFOS,
@@ -399,7 +401,9 @@ export function uniswapV2LikeRouterAddress(
         | ERC20BridgeSource.ComethSwap
         | ERC20BridgeSource.Dfyn
         | ERC20BridgeSource.WaultSwap
-        | ERC20BridgeSource.Polydex,
+        | ERC20BridgeSource.Polydex
+        | ERC20BridgeSource.ShibaSwap
+        | ERC20BridgeSource.JetSwap,
 ): string {
     switch (source) {
         case ERC20BridgeSource.UniswapV2:
@@ -432,6 +436,10 @@ export function uniswapV2LikeRouterAddress(
             return WAULTSWAP_ROUTER_BY_CHAIN_ID[chainId];
         case ERC20BridgeSource.Polydex:
             return POLYDEX_ROUTER_BY_CHAIN_ID[chainId];
+        case ERC20BridgeSource.ShibaSwap:
+            return SHIBASWAP_ROUTER_BY_CHAIN_ID[chainId];
+        case ERC20BridgeSource.JetSwap:
+            return JETSWAP_ROUTER_BY_CHAIN_ID[chainId];
         default:
             throw new Error(`Unknown UniswapV2 like source ${source}`);
     }