+++ packages/asset-swapper/src/utils/market_operation_utils/path.ts
@@ -78,6 +78,11 @@ export class Path {
         return this;
     }
 
+    /**
+     * Add a fallback path to the current path
+     * Fallback must contain exclusive fills that are
+     * not present in this path
+     */
     public addFallback(fallback: Path): this {
         // If the last fill is Native and penultimate is not, then the intention was to partial fill
         // In this case we drop it entirely as we can't handle a failure at the end and we don't
@@ -93,7 +98,16 @@ export class Path {
         // an additional protocol fee. I.e [Uniswap,Native,Kyber] becomes [Native,Uniswap,Kyber]
         // In the previous step we dropped any hanging Native partial fills, as to not fully fill
         const nativeFills = this.fills.filter(f => f.source === ERC20BridgeSource.Native);
-        this.fills = [...nativeFills.filter(f => f !== lastNativeFillIfExists), ...fallback.fills];
+        const otherFills = this.fills.filter(f => f.source !== ERC20BridgeSource.Native);
+        const otherSourcePathIds = otherFills.map(f => f.sourcePathId);
+        this.fills = [
+            // Append all of the native fills first
+            ...nativeFills.filter(f => f !== lastNativeFillIfExists),
+            // Add the other fills that are not native in the optimal path
+            ...otherFills,
+            // Add the fallbacks to the end that aren't already included
+            ...fallback.fills.filter(f => !otherSourcePathIds.includes(f.sourcePathId)),
+        ];
         // Recompute the source flags
         this.sourceFlags = this.fills.reduce((flags, fill) => flags | fill.flags, BigInt(0));
         return this;
@@ -101,7 +115,7 @@ export class Path {
 
     public collapse(opts: CreateOrderFromPathOpts): CollapsedPath {
         const [makerToken, takerToken] = getMakerTakerTokens(opts);
-        const collapsedFills = this.collapsedFills === undefined ? this._collapseFills() : this.collapsedFills;
+        const collapsedFills = this.collapsedFills === undefined ? this.collapseFills() : this.collapsedFills;
         this.orders = [];
         for (let i = 0; i < collapsedFills.length; ) {
             if (collapsedFills[i].source === ERC20BridgeSource.Native) {
@@ -230,7 +244,7 @@ export class Path {
         return true;
     }
 
-    private _collapseFills(): ReadonlyArray<CollapsedFill> {
+    public collapseFills(): ReadonlyArray<CollapsedFill> {
         this.collapsedFills = [];
         for (const fill of this.fills) {
             const source = fill.source;