+++ packages/asset-swapper/src/utils/market_operation_utils/types.ts
@@ -68,6 +68,7 @@ export enum ERC20BridgeSource {
     UniswapV3 = 'Uniswap_V3',
     CurveV2 = 'Curve_V2',
     Lido = 'Lido',
+    ShibaSwap = 'ShibaSwap',
     // BSC only
     PancakeSwap = 'PancakeSwap',
     PancakeSwapV2 = 'PancakeSwap_V2',
@@ -86,6 +87,7 @@ export enum ERC20BridgeSource {
     WaultSwap = 'WaultSwap',
     Polydex = 'Polydex',
     FirebirdOneSwap = 'FirebirdOneSwap',
+    JetSwap = 'JetSwap',
 }
 export type SourcesWithPoolsCache = ERC20BridgeSource.Balancer | ERC20BridgeSource.BalancerV2 | ERC20BridgeSource.Cream;